fix: widen the hub's public stream parameter to intptr_t + bump cpu/cuda pins
A `cudaStream_t` is a pointer and therefore 64-bit, but every public
`stream` parameter in the hub's device-dispatch functions was declared
`int`. `ff::dt_euclidean` and friends inspect the DLTensor's device and
forward the same argument to either `FF_CPU::` or `FF_CUDA::`:

    if (is_cuda) return FF_CUDA::dt_euclidean(inp_out, voxel_spacing, stream);
    else         return FF_CPU::dt_euclidean(inp_out, voxel_spacing, stream);

so both backends must agree on the parameter's type. Widens 114 parameter
declarations across the 8 modules to `intptr_t`, matching the now-widened
cpu/cuda dispatch layers. This is a signature-only change (verified with the
same non-signature-line grep used in the cpu-lib/cuda-lib PRs: empty).

Bumps `cpu` -> d969e2d (fastfields-cpu-lib#61) and `cuda` -> 33aa9b3
(fastfields-cuda-lib#35).

Verified with `make all CXX=clang++`: the full hub (cpu library + hub object
files + link) builds successfully end to end against the bumped pins,
producing libfastfields.so linked against libfastfields-cpu.so. This is a
real, complete build of the CPU path; the CUDA path is not linked by this
target (CUDA needs `USE_CUDA=1` and nvcc, exercised at the cuda-lib layer
instead).

Refs fastfields/fastfields-lib#4

+++ reg_field.h
@@ -1,6 +1,7 @@
 #ifndef FF_LIB_REG_FIELD
 #define FF_LIB_REG_FIELD
 #include "dlpack.h"
+#include <cstdint>
 #include "defines.h"
 
 #ifndef FF_LIB_BOUND_SPLINE_T
@@ -67,7 +68,7 @@ void field_matvec(
     const double   * bending   = nullptr,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -83,7 +84,7 @@ void field_addmatvec_(
     const double   * bending   = nullptr,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -99,7 +100,7 @@ void field_submatvec_(
     const double   * bending   = nullptr,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -114,7 +115,7 @@ void field_diag(
     const double   * bending   = nullptr,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -135,7 +136,7 @@ void field_kernel(
     const double   * bending   = nullptr,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -165,7 +166,7 @@ void field_adddiag_(
     const double   * bending   = nullptr,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -181,7 +182,7 @@ void field_subdiag_(
     const double   * bending   = nullptr,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -197,7 +198,7 @@ void field_addkernel_(
     const double   * bending   = nullptr,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -213,7 +214,7 @@ void field_subkernel_(
     const double   * bending   = nullptr,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 void field_relax(
@@ -227,7 +228,7 @@ void field_relax(
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
           int        nb_iter   = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -251,7 +252,7 @@ void field_forward(
     const double   * bending   = nullptr,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -274,7 +275,7 @@ void field_precond(
     const double   * bending   = nullptr,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -293,7 +294,7 @@ void field_precond_(
     const double   * bending   = nullptr,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -325,7 +326,7 @@ void field_matvec_rls(
     const double   * bending   = nullptr,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -341,7 +342,7 @@ void field_diag_rls(
     const double   * bending   = nullptr,
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -366,7 +367,7 @@ void field_relax_rls(
           int8_t     bound     = bound_t::DCT2,
           int        ndim      = 1,
           int        nb_iter   = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 FF_NAMESPACE_END(FF)
